browniedir/requirements.txt
eth-brownie==1.20.6
web3
aiofiles
packaging
solc-select
python-dotenv